MITRE refactor phase #2: Moving mitre utils to pkg
