+++ pkg/mitre/utils/utils.go
